+++ dojo/api_v2/views.py
@@ -1,7 +1,5 @@
 import logging
-import mimetypes
 from datetime import datetime
-from pathlib import Path
 
 import pghistory
 from dateutil.relativedelta import relativedelta
@@ -10,8 +8,6 @@ from django.contrib.auth.models import Permission
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError
 from django.db.models.query import QuerySet as DjangoQuerySet
-from django.http import FileResponse
-from django.urls import reverse
 from django.utils import timezone
 from django_filters.rest_framework import DjangoFilterBackend
 from drf_spectacular.renderers import OpenApiJsonRenderer2
@@ -41,7 +37,6 @@ from dojo.endpoint.ui.views import get_endpoint_ids
 from dojo.filters import (
     ApiAppAnalysisFilter,
     ApiDojoMetaFilter,
-    ApiRiskAcceptanceFilter,
 )
 from dojo.finding.ui.filters import (
     ReportFindingFilter,
@@ -61,11 +56,8 @@ from dojo.models import (
     Language_Type,
     Languages,
     Network_Locations,
-    NoteHistory,
-    Notes,
     Product,
     Regulation,
-    Risk_Acceptance,
     SLA_Configuration,
     Sonarqube_Issue,
     Sonarqube_Issue_Transition,
@@ -82,8 +74,6 @@ from dojo.reports.ui.views import (
     prefetch_related_findings_for_report,
     report_url_resolver,
 )
-from dojo.risk_acceptance.helper import remove_finding_from_risk_acceptance
-from dojo.risk_acceptance.queries import get_authorized_risk_acceptances
 from dojo.test.queries import get_authorized_tests
 from dojo.user.utils import get_configuration_permissions_codenames
 from dojo.utils import (
@@ -91,7 +81,6 @@ from dojo.utils import (
     get_celery_queue_length,
     get_celery_worker_status,
     get_system_setting,
-    process_tag_notifications,
     purge_celery_queue,
     purge_celery_queue_by_task_name,
 )
@@ -170,128 +159,6 @@ class DeprecationNoticeMixin:
 
 # @extend_schema_view(**schema_with_prefetch())
 # Nested models with prefetch make the response schema too long for Swagger UI
-class RiskAcceptanceViewSet(
-    PrefetchDojoModelViewSet,
-):
-    serializer_class = serializers.RiskAcceptanceSerializer
-    queryset = Risk_Acceptance.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_class = ApiRiskAcceptanceFilter
-
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasRiskAcceptancePermission,
-    )
-
-    def destroy(self, request, pk=None):
-        instance = self.get_object()
-        # Remove any findings on the risk acceptance
-        for finding in instance.accepted_findings.all():
-            remove_finding_from_risk_acceptance(request.user, instance, finding)
-        # return the response of the object being deleted
-        return super().destroy(request, pk=pk)
-
-    def get_queryset(self):
-        return (
-            get_authorized_risk_acceptances("edit")
-            .prefetch_related(
-                "notes", "engagement_set", "owner", "accepted_findings",
-            )
-            .distinct()
-        )
-
-    @extend_schema(
-        methods=["GET"],
-        responses={
-            status.HTTP_200_OK: serializers.RiskAcceptanceToNotesSerializer,
-        },
-    )
-    @extend_schema(
-        methods=["POST"],
-        request=serializers.AddNewNoteOptionSerializer,
-        responses={status.HTTP_201_CREATED: serializers.NoteSerializer},
-    )
-    @action(detail=True, methods=["get", "post"], permission_classes=(IsAuthenticated, permissions.UserHasRiskAcceptanceRelatedObjectPermission))
-    def notes(self, request, pk=None):
-        risk_acceptance = self.get_object()
-        if request.method == "POST":
-            new_note = serializers.AddNewNoteOptionSerializer(data=request.data)
-            if new_note.is_valid():
-                entry = new_note.validated_data["entry"]
-                private = new_note.validated_data.get("private", False)
-                note_type = new_note.validated_data.get("note_type", None)
-            else:
-                return Response(new_note.errors, status=status.HTTP_400_BAD_REQUEST)
-
-            notes = risk_acceptance.notes.filter(note_type=note_type).first()
-            if notes and note_type and note_type.is_single:
-                return Response("Only one instance of this note_type allowed on a risk acceptance.", status=status.HTTP_400_BAD_REQUEST)
-
-            author = request.user
-            note = Notes(entry=entry, author=author, private=private, note_type=note_type)
-            note.save()
-            history = NoteHistory.objects.create(data=note.entry, time=note.date, current_editor=note.author)
-            note.history.add(history)
-            risk_acceptance.notes.add(note)
-            engagement = risk_acceptance.engagement
-            if engagement:
-                process_tag_notifications(
-                    request=request,
-                    note=note,
-                    parent_url=request.build_absolute_uri(
-                        reverse("view_risk_acceptance", args=(engagement.id, risk_acceptance.id)),
-                    ),
-                    parent_title=f"Risk Acceptance: {risk_acceptance.name}",
-                )
-
-            serialized_note = serializers.NoteSerializer(
-                {"author": author, "entry": entry, "private": private},
-            )
-            return Response(serialized_note.data, status=status.HTTP_201_CREATED)
-
-        notes = risk_acceptance.notes.all()
-        serialized_notes = serializers.RiskAcceptanceToNotesSerializer(
-            {"risk_acceptance_id": risk_acceptance, "notes": notes},
-        )
-        return Response(serialized_notes.data, status=status.HTTP_200_OK)
-
-    @extend_schema(
-        methods=["GET"],
-        responses={
-            status.HTTP_200_OK: serializers.RiskAcceptanceProofSerializer,
-        },
-    )
-    @action(detail=True, methods=["get"], permission_classes=(IsAuthenticated, permissions.UserHasRiskAcceptanceRelatedObjectPermission))
-    def download_proof(self, request, pk=None):
-        risk_acceptance = self.get_object()
-        # Get the file object
-        file_object = risk_acceptance.path
-        if file_object is None or risk_acceptance.filename() is None:
-            return Response(
-                {"error": "Proof has not provided to this risk acceptance..."},
-                status=status.HTTP_404_NOT_FOUND,
-            )
-        # Get the path of the file in media root
-        file_path = Path(settings.MEDIA_ROOT) / file_object.name
-        # NOTE: FileResponse takes ownership of closing the file handle when the response is closed.
-        # Explicitly register the closer to avoid potential resource leaks and satisfy static analyzers.
-        file_handle = file_path.open("rb")
-        # send file
-        response = FileResponse(
-            file_handle,
-            content_type=mimetypes.guess_type(str(file_path))[0] or "application/octet-stream",
-            status=status.HTTP_200_OK,
-        )
-        if hasattr(response, "_resource_closers"):
-            response._resource_closers.append(file_handle.close)
-        response["Content-Length"] = file_object.size
-        response[
-            "Content-Disposition"
-        ] = f'attachment; filename="{risk_acceptance.filename()}"'
-
-        return response
-
-
 # These are technologies in the UI and the API!
 # Authorization: object-based
 @extend_schema_view(**schema_with_prefetch())
