+++ dojo/filters.py
@@ -51,7 +51,6 @@ from dojo.models import (
     Note_Type,
     Product,
     Product_Type,
-    Risk_Acceptance,
     Test,
     Vulnerability_Id,
 )
@@ -1271,41 +1270,6 @@ class MetricsEndpointFilterWithoutObjectLookups(MetricsEndpointFilterHelper, Fin
     class Meta:
         model = Endpoint_Status
         exclude = ["last_modified", "endpoint", "finding"]
-
-
-class ApiRiskAcceptanceFilter(DojoFilter):
-    created = DateRangeFilter()
-    updated = DateRangeFilter()
-
-    o = OrderingFilter(
-        # tuple-mapping retains order
-        fields=(
-            ("name", "name"),
-            ("created", "created"),
-            ("updated", "updated"),
-        ),
-    )
-
-    class Meta:
-        model = Risk_Acceptance
-        fields = {
-            "name": ["exact", "icontains"],
-            "accepted_findings": ["exact"],
-            "recommendation": ["exact"],
-            "recommendation_details": ["exact", "icontains"],
-            "decision": ["exact"],
-            "decision_details": ["exact", "icontains"],
-            "accepted_by": ["exact", "icontains"],
-            "owner": ["exact"],
-            "expiration_date": ["exact", "gt", "lt", "gte", "lte"],
-            "expiration_date_warned": ["exact", "gt", "lt", "gte", "lte"],
-            "expiration_date_handled": ["exact", "gt", "lt", "gte", "lte"],
-            "reactivate_expired": ["exact"],
-            "restart_sla_expired": ["exact"],
-            "notes": ["exact"],
-            "created": ["exact", "gt", "lt", "gte", "lte"],
-            "updated": ["exact", "gt", "lt", "gte", "lte"],
-        }
 
 
 class ApiAppAnalysisFilter(DojoFilter):
