+++ dojo/risk_acceptance/__init__.py
@@ -1,0 +1,1 @@
+import dojo.risk_acceptance.admin  # noqa: F401
