+++ dojo/risk_acceptance/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.risk_acceptance.models import Risk_Acceptance
+
+admin.site.register(Risk_Acceptance)
